move the gyacomo config routine into the gyacomo interface
